split up auth controller
