+++ internal/services/gateway/auth_controller_test.go
@@ -14,29 +14,34 @@
 package gateway
 
 import (
-	"bytes"
-	"context"
-	"encoding/json"
 	"net/http"
 	"net/http/httptest"
 	"os"
 	"path/filepath"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/g8e-ai/g8e/internal/config"
-	"github.com/g8e-ai/g8e/internal/constants"
-	"github.com/g8e-ai/g8e/internal/models"
 	"github.com/g8e-ai/g8e/internal/response"
 	"github.com/g8e-ai/g8e/internal/services/keystore"
 	"github.com/g8e-ai/g8e/internal/services/mcp"
 	"github.com/g8e-ai/g8e/internal/services/storage"
 	"github.com/g8e-ai/g8e/internal/testutil"
 )
+
+// mockActuatorKeyReader is a test implementation of actuatorKeyReader.
+type mockActuatorKeyReader struct {
+	keyID     string
+	publicKey string
+	err       error
+}
+
+func (m *mockActuatorKeyReader) ReadActuatorPublicKey() (keyID, publicKey string, err error) {
+	return m.keyID, m.publicKey, m.err
+}
 
 func setupTestAuthController(t *testing.T) (*AuthController, *config.Config) {
 	t.Helper()
@@ -105,2084 +110,122 @@ func setupTestAuthController(t *testing.T) (*AuthController, *config.Config) {
 		t.Fatalf("failed to create MCP gateway: %v", err)
 	}
 
-	authController := newAuthController(cfg, logger, db, auth, passkey, userSvc, reg, pki, webSessionSvc, cliSessionSvc, operatorSessionSvc, suspendedTxService, mcpGateway, resp)
+	authController := newAuthController(cfg, logger, db, auth, passkey, userSvc, reg, pki, webSessionSvc, cliSessionSvc, operatorSessionSvc, suspendedTxService, mcpGateway, resp, nil)
 	return authController, cfg
 }
 
-func TestHandleBootstrap(t *testing.T) {
-	t.Run("Success - Bootstrap with CSR over loopback", func(t *testing.T) {
+// Test Helpers
+
+func testMethodNotAllowed(t *testing.T, handler http.HandlerFunc, method, url string) {
+	t.Helper()
+	req := httptest.NewRequest(method, url, nil)
+	rr := httptest.NewRecorder()
+	handler.ServeHTTP(rr, req)
+	assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
+}
+
+func testInvalidJSON(t *testing.T, handler http.HandlerFunc, method, url string) {
+	t.Helper()
+	req := httptest.NewRequest(method, url, strings.NewReader("{invalid}"))
+	rr := httptest.NewRecorder()
+	handler.ServeHTTP(rr, req)
+	assert.Equal(t, http.StatusBadRequest, rr.Code)
+}
+
+func testMissingUserID(t *testing.T, handler http.HandlerFunc, method, url string) {
+	t.Helper()
+	req := httptest.NewRequest(method, url, strings.NewReader(`{"not_user_id":"data"}`))
+	rr := httptest.NewRecorder()
+	handler.ServeHTTP(rr, req)
+	assert.Equal(t, http.StatusBadRequest, rr.Code)
+	assert.Contains(t, rr.Body.String(), "user_id required")
+}
+
+func TestAuthControllerReadBody(t *testing.T) {
+	t.Run("Success - reads valid JSON body", func(t *testing.T) {
 		t.Parallel()
 		c, _ := setupTestAuthController(t)
-		csr := testutil.GenerateTestCSRP256(t, "test-operator")
-		cliCsr := testutil.GenerateTestCSRP256(t, "test-cli")
-		body := map[string]string{
-			"name":               "Owner",
-			"csr_pem":            csr,
-			"cli_csr_pem":        cliCsr,
-			"system_fingerprint": "test-fp",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
+		body := `{"test":"data"}`
+		req := httptest.NewRequest(http.MethodPost, "/test", strings.NewReader(body))
 
-		c.handleLocalBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusCreated, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
+		data, err := c.readBody(req)
 		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["operator_cert"])
-		assert.NotEmpty(t, resp["operator_cert_chain"])
-		assert.NotEmpty(t, resp["hub_trust_bundle"])
-		assert.NotEmpty(t, resp["operator_session_id"])
-		assert.NotEmpty(t, resp["cli_session_id"])
-		assert.NotEqual(t, resp["operator_session_id"], resp["cli_session_id"],
-			"cli_session_id MUST be a distinct identifier from operator_session_id")
+		assert.Equal(t, []byte(body), data)
 	})
 
-	t.Run("Failure - Non-loopback CSR request rejected", func(t *testing.T) {
+	t.Run("Success - reads empty body", func(t *testing.T) {
 		t.Parallel()
 		c, _ := setupTestAuthController(t)
-		csr := testutil.GenerateTestCSRP256(t, "test-operator")
-		cliCsr := testutil.GenerateTestCSRP256(t, "test-cli")
-		body := map[string]string{
-			"name":               "Owner",
-			"csr_pem":            csr,
-			"cli_csr_pem":        cliCsr,
-			"system_fingerprint": "test-fp",
-		}
-		b, err := json.Marshal(body)
+		req := httptest.NewRequest(http.MethodPost, "/test", strings.NewReader(""))
+
+		data, err := c.readBody(req)
 		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		req.RemoteAddr = "192.168.1.1:12345"
-		rr := httptest.NewRecorder()
-
-		c.handleLocalBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.JSONEq(t, `{"error":"CSR auto-issue only available over loopback"}`, rr.Body.String())
+		assert.Equal(t, []byte{}, data)
 	})
 
-	t.Run("Success - Rotation for existing bootstrap user", func(t *testing.T) {
+	t.Run("Failure - body exceeds max payload bytes", func(t *testing.T) {
 		t.Parallel()
 		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateBootstrapUser()
-		require.NoError(t, err)
-		require.NotNil(t, user)
+		// Set a small max payload for testing
+		c.cfg.Gateway.MaxPayloadBytes = 100
+		largeBody := strings.Repeat("a", 200)
+		req := httptest.NewRequest(http.MethodPost, "/test", strings.NewReader(largeBody))
 
-		csr := testutil.GenerateTestCSRP256(t, "test-operator")
-		cliCsr := testutil.GenerateTestCSRP256(t, "test-cli")
-		body := map[string]string{
-			"name":               "Owner",
-			"csr_pem":            csr,
-			"cli_csr_pem":        cliCsr,
-			"system_fingerprint": "rotated-fp",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		c.handleLocalBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusCreated, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["operator_cert"])
-	})
-
-	t.Run("Failure - Rotation fails for disabled bootstrap user", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, _ := c.userSvc.CreateBootstrapUser()
-		c.userSvc.Disable(user.ID, "retired", "actor", "op")
-
-		csr := testutil.GenerateTestCSRP256(t, "test-operator")
-		cliCsr := testutil.GenerateTestCSRP256(t, "test-cli")
-		body := map[string]string{
-			"name":               "Owner",
-			"csr_pem":            csr,
-			"cli_csr_pem":        cliCsr,
-			"system_fingerprint": "fail-fp",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		c.handleLocalBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusConflict, rr.Code)
-		assert.JSONEq(t, `{"error":"bootstrap user is disabled, cannot rotate"}`, rr.Body.String())
-	})
-
-	t.Run("Failure - Rejects bootstrap if ANY other users exist", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		c.userSvc.CreateUser()
-
-		body := map[string]string{
-			"name": "Superadmin",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleLocalBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.JSONEq(t, `{"error":"bootstrap only available for initial setup"}`, rr.Body.String())
+		_, err := c.readBody(req)
+		assert.Error(t, err)
+		assert.Contains(t, err.Error(), "http: request body too large")
 	})
 }
 
-func TestHandleBootstrapStatus(t *testing.T) {
-	t.Run("Initially not bootstrapped", func(t *testing.T) {
+func TestFileActuatorKeyReader(t *testing.T) {
+	t.Run("Success - reads valid actuator key file", func(t *testing.T) {
 		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/auth/bootstrap/status", nil)
-		rr := httptest.NewRecorder()
-		c.handleBootstrapStatus(rr, req)
+		// Create a temporary file with valid actuator key data
+		tmpDir := t.TempDir()
+		keyFile := filepath.Join(tmpDir, "actuator_key.json")
+		keyData := `{"key_id":"test-key-id","public_key":"test-public-key"}`
+		require.NoError(t, os.WriteFile(keyFile, []byte(keyData), 0644))
 
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err := json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.False(t, resp["bootstrapped"].(bool))
-	})
-
-	t.Run("Bootstrapped after creating a user", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		_, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/auth/bootstrap/status", nil)
-		rr := httptest.NewRecorder()
-		c.handleBootstrapStatus(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["bootstrapped"].(bool))
-	})
-}
-
-func TestHandleAuthPasskeysRegisterChallenge(t *testing.T) {
-	t.Run("Success - valid request", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id":   user.ID,
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["options"])
-	})
-
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/register/challenge", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/register/challenge", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Success - JIT route with first credential", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id":   user.ID,
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/jit-register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-	})
-
-	t.Run("Failure - JIT route with existing credentials", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		// Add a fake credential to simulate existing credentials
-		user.PasskeyCredentials = []models.PasskeyCredential{{ID: []byte("existing-cred")}}
-		updatedUser, err := json.Marshal(user)
-		require.NoError(t, err)
-		c.db.DocStore.DocSet("users", user.ID, updatedUser)
-
-		body := map[string]string{
-			"user_id":   user.ID,
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/jit-register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "first-credential registration only")
-	})
-
-	t.Run("Success - session context user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id":   "",
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/register/challenge", bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-	})
-
-	t.Run("Failure - session user_id mismatch", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id":   "other-user-id",
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/register/challenge", bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id mismatch")
-	})
-}
-
-func TestHandleAuthPasskeysRegisterVerify(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/register/verify", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/register/verify", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/register/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Failure - JIT route with existing credentials", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		// Add a fake credential to simulate existing credentials
-		user.PasskeyCredentials = []models.PasskeyCredential{{ID: []byte("existing-cred")}}
-		updatedUser, err := json.Marshal(user)
-		require.NoError(t, err)
-		c.db.DocStore.DocSet("users", user.ID, updatedUser)
-
-		body := map[string]string{
-			"user_id": user.ID,
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/jit-register/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "first-credential registration only")
-	})
-}
-
-func TestHandleAuthPasskeysAuthenticateChallenge(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/authenticate/challenge", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/authenticate/challenge", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/authenticate/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Failure - no passkeys registered", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id": user.ID,
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/authenticate/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.False(t, resp["success"].(bool))
-		assert.Contains(t, resp["error"].(string), "no passkeys registered")
-	})
-
-	t.Run("Success - session context user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id": "",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/authenticate/challenge", bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-	})
-
-	t.Run("Failure - session user_id mismatch", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id": "other-user-id",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/authenticate/challenge", bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id mismatch")
-	})
-}
-
-func TestHandleAuthPasskeysAuthenticateVerify(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/authenticate/verify", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateVerify(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/authenticate/verify", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/authenticate/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Success - session context user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id": "",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/authenticate/verify", bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateVerify(rr, req)
-
-		// Will fail verification since no real assertion response, but should get past validation
-		assert.Equal(t, http.StatusOK, rr.Code)
-	})
-
-	t.Run("Failure - session user_id mismatch", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id": "other-user-id",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/authenticate/verify", bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysAuthenticateVerify(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id mismatch")
-	})
-}
-
-func TestHandleAuthPasskeys(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeys(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeys(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Success - list credentials", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys?user_id="+user.ID, nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeys(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		// Credentials may be nil or empty slice when no credentials exist
-		creds, ok := resp["credentials"]
-		assert.True(t, ok)
-		if creds != nil {
-			assert.IsType(t, []interface{}{}, creds)
-		}
-	})
-}
-
-func TestHandleAuthPasskeysRevoke(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/cred-id", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRevoke(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodDelete, "/api/v1/auth/passkeys/cred-id", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRevoke(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Failure - missing credential_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodDelete, "/api/v1/auth/passkeys/?user_id="+user.ID, nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRevoke(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "credential_id required")
-	})
-
-	t.Run("Success - revoke credential", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodDelete, "/api/v1/auth/passkeys/test-cred-id?user_id="+user.ID, nil)
-		rr := httptest.NewRecorder()
-
-		c.handleAuthPasskeysRevoke(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.False(t, resp["found"].(bool)) // No credential exists
-	})
-}
-
-func TestHandleApprovalAction(t *testing.T) {
-	t.Run("Failure - unauthorized", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approvals/txhash123", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalAction(rr, req)
-
-		assert.Equal(t, http.StatusUnauthorized, rr.Code)
-		assert.Contains(t, rr.Body.String(), "unauthorized")
-	})
-
-	t.Run("Failure - transaction not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approvals/nonexistent-tx", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalAction(rr, req)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "transaction not found")
-	})
-}
-
-func TestHandleApprovalChallenge(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approvals/txhash123/challenge", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalChallenge(rr, req, "txhash123", user.ID)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - transaction not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/approvals/nonexistent/challenge", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalChallenge(rr, req, "nonexistent", user.ID)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "transaction not found")
-	})
-
-	t.Run("Failure - transaction belongs to another user", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user1, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-		user2, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		// Create a suspended transaction for user1 via suspended store
-		txHash := "txhash123"
-		suspendedTx := &models.SuspendedTransaction{
-			TransactionHash: txHash,
-			UserID:          user1.ID,
-			ToolName:        "test-tool",
-			ToolArguments:   []byte("{}"),
-			ExpiresAt:       time.Now().Add(5 * time.Minute),
-		}
-		c.suspendedStore.StoreSuspendedTransaction(context.Background(), suspendedTx)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/approvals/"+txHash+"/challenge", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user2.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalChallenge(rr, req, txHash, user2.ID)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "transaction belongs to another user")
-	})
-}
-
-func TestHandleApprovalVerify(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/approvals/txhash123/verify", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalVerify(rr, req, "txhash123", user.ID)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - transaction not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approvals/nonexistent/verify", strings.NewReader("{}"))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalVerify(rr, req, "nonexistent", user.ID)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "transaction not found")
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		txHash := "txhash123"
-		suspendedTx := &models.SuspendedTransaction{
-			TransactionHash: txHash,
-			UserID:          user.ID,
-			ToolName:        "test-tool",
-			ToolArguments:   []byte("{}"),
-			ExpiresAt:       time.Now().Add(5 * time.Minute),
-		}
-		c.suspendedStore.StoreSuspendedTransaction(context.Background(), suspendedTx)
-
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approvals/"+txHash+"/verify", strings.NewReader("{invalid}"))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalVerify(rr, req, txHash, user.ID)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-}
-
-func TestHandleCLIApproval(t *testing.T) {
-	t.Run("Failure - transaction not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"cli_signature":         "sig123",
-			"mtls_cert_fingerprint": "fp123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approvals/nonexistent", bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIApproval(rr, req, "nonexistent", user.ID)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "transaction not found")
-	})
-
-	t.Run("Failure - missing cli_signature", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		txHash := "txhash123"
-		suspendedTx := &models.SuspendedTransaction{
-			TransactionHash: txHash,
-			UserID:          user.ID,
-			ToolName:        "test-tool",
-			ToolArguments:   []byte("{}"),
-			ExpiresAt:       time.Now().Add(5 * time.Minute),
-		}
-		c.suspendedStore.StoreSuspendedTransaction(context.Background(), suspendedTx)
-
-		body := map[string]string{
-			"mtls_cert_fingerprint": "fp123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approvals/"+txHash, bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIApproval(rr, req, txHash, user.ID)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "cli_signature required")
-	})
-
-	t.Run("Failure - missing mtls_cert_fingerprint", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		txHash := "txhash123"
-		suspendedTx := &models.SuspendedTransaction{
-			TransactionHash: txHash,
-			UserID:          user.ID,
-			ToolName:        "test-tool",
-			ToolArguments:   []byte("{}"),
-			ExpiresAt:       time.Now().Add(5 * time.Minute),
-		}
-		c.suspendedStore.StoreSuspendedTransaction(context.Background(), suspendedTx)
-
-		body := map[string]string{
-			"cli_signature": "sig123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approvals/"+txHash, bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIApproval(rr, req, txHash, user.ID)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "mtls_cert_fingerprint required")
-	})
-}
-
-func TestHandleCLIEnrollment(t *testing.T) {
-	t.Run("Success - CLI enrollment over loopback after bootstrap", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		bootstrapUser, err := c.userSvc.CreateBootstrapUser()
-		require.NoError(t, err)
-		require.NotNil(t, bootstrapUser)
-
-		cliCSR := testutil.GenerateTestCSRP256(t, "test-cli")
-		body := map[string]string{
-			"cli_csr_pem":        cliCSR,
-			"system_fingerprint": "test-fp",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/cli/enroll", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		c.handleCLIEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusCreated, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["cli_cert"])
-		assert.NotEmpty(t, resp["cli_cert_chain"])
-		assert.NotEmpty(t, resp["cli_session_id"])
-		assert.NotEmpty(t, resp["user_id"])
-		assert.NotEmpty(t, resp["hub_trust_bundle"])
-		assert.NotEmpty(t, resp["operator_session_id"])
-		assert.NotEmpty(t, resp["operator_id"])
-	})
-
-	t.Run("Failure - Non-loopback request rejected", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		bootstrapUser, err := c.userSvc.CreateBootstrapUser()
-		require.NoError(t, err)
-		require.NotNil(t, bootstrapUser)
-
-		cliCSR := testutil.GenerateTestCSRP256(t, "test-cli")
-		body := map[string]string{
-			"cli_csr_pem":        cliCSR,
-			"system_fingerprint": "test-fp",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/cli/enroll", bytes.NewReader(b))
-		req.RemoteAddr = "192.168.1.1:12345"
-		rr := httptest.NewRecorder()
-
-		c.handleCLIEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "CLI enrollment only available over loopback")
-	})
-
-	t.Run("Failure - Rejected when not bootstrapped", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-
-		cliCSR := testutil.GenerateTestCSRP256(t, "test-cli")
-		body := map[string]string{
-			"cli_csr_pem":        cliCSR,
-			"system_fingerprint": "test-fp",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/cli/enroll", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		c.handleCLIEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "CLI enrollment only available after bootstrap")
-	})
-
-	t.Run("Failure - Rejected when bootstrap user disabled", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		bootstrapUser, err := c.userSvc.CreateBootstrapUser()
-		require.NoError(t, err)
-		c.userSvc.Disable(bootstrapUser.ID, "retired", "actor", "op")
-
-		cliCSR := testutil.GenerateTestCSRP256(t, "test-cli")
-		body := map[string]string{
-			"cli_csr_pem":        cliCSR,
-			"system_fingerprint": "test-fp",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/cli/enroll", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		c.handleCLIEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusConflict, rr.Code)
-		assert.Contains(t, rr.Body.String(), "bootstrap user is disabled")
-	})
-
-	t.Run("Failure - Missing cli_csr_pem", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		bootstrapUser, err := c.userSvc.CreateBootstrapUser()
-		require.NoError(t, err)
-		require.NotNil(t, bootstrapUser)
-
-		body := map[string]string{
-			"system_fingerprint": "test-fp",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/cli/enroll", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		c.handleCLIEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "cli_csr_pem is required")
-	})
-
-	t.Run("Failure - Method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		bootstrapUser, err := c.userSvc.CreateBootstrapUser()
-		require.NoError(t, err)
-		require.NotNil(t, bootstrapUser)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/cli/enroll", nil)
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		c.handleCLIEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-}
-
-func TestHandleApprovalPage(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approve/txhash123", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalPage(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - missing transaction hash", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/approve/", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalPage(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "transaction hash required")
-	})
-
-	t.Run("Failure - transaction not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/approve/nonexistent", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalPage(rr, req)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "transaction not found")
-	})
-
-	t.Run("Success - returns HTML page", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		txHash := "txhash123"
-		suspendedTx := &models.SuspendedTransaction{
-			TransactionHash: txHash,
-			UserID:          user.ID,
-			ToolName:        "test-tool",
-			ToolArguments:   []byte(`{"arg":"value"}`),
-			ExpiresAt:       time.Now().Add(5 * time.Minute),
-		}
-		c.suspendedStore.StoreSuspendedTransaction(context.Background(), suspendedTx)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/approve/"+txHash, nil)
-		rr := httptest.NewRecorder()
-
-		c.handleApprovalPage(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		assert.Equal(t, "text/html; charset=utf-8", rr.Header().Get("Content-Type"))
-		assert.Contains(t, rr.Body.String(), "Approve Transaction")
-		assert.Contains(t, rr.Body.String(), txHash)
-		assert.Contains(t, rr.Body.String(), "test-tool")
-	})
-}
-
-func TestHandleListSuspendedTransactions(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/approvals", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleListSuspendedTransactions(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - unauthorized", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/approvals", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleListSuspendedTransactions(rr, req)
-
-		assert.Equal(t, http.StatusUnauthorized, rr.Code)
-		assert.Contains(t, rr.Body.String(), "unauthorized")
-	})
-
-	t.Run("Success - empty list", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/approvals", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleListSuspendedTransactions(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		// When empty, transactions may be null or empty array
-		transactions, ok := resp["transactions"].([]interface{})
-		if !ok {
-			// If null, that's acceptable for empty list
-			assert.Nil(t, resp["transactions"])
-		} else {
-			assert.Empty(t, transactions)
-		}
-	})
-
-	t.Run("Success - with query user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/approvals?user_id="+user.ID, nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleListSuspendedTransactions(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		// When empty, transactions may be null or empty array
-		transactions, ok := resp["transactions"].([]interface{})
-		if !ok {
-			// If null, that's acceptable for empty list
-			assert.Nil(t, resp["transactions"])
-		} else {
-			assert.Empty(t, transactions)
-		}
-	})
-}
-
-func TestHandleUserMe(t *testing.T) {
-	t.Run("Failure - missing user_id in context", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/me", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleUserMe(rr, req)
-
-		assert.Equal(t, http.StatusUnauthorized, rr.Code)
-		assert.Contains(t, rr.Body.String(), "unauthorized")
-	})
-
-	t.Run("Success - returns user data", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/me", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleUserMe(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotNil(t, resp["user"])
-	})
-
-	t.Run("Failure - user not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/me", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, "nonexistent-user"))
-		rr := httptest.NewRecorder()
-
-		c.handleUserMe(rr, req)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user not found")
-	})
-}
-
-func TestHandleWebSession(t *testing.T) {
-	t.Run("Failure - missing user_id in context", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/websession", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleWebSession(rr, req)
-
-		assert.Equal(t, http.StatusUnauthorized, rr.Code)
-		assert.Contains(t, rr.Body.String(), "unauthorized")
-	})
-
-	t.Run("Success - returns session data with cookie", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/websession", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		req.AddCookie(&http.Cookie{Name: "g8e_session", Value: "test-session-id"})
-		rr := httptest.NewRecorder()
-
-		c.handleWebSession(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.Equal(t, user.ID, resp["user_id"])
-		assert.Equal(t, "test-session-id", resp["web_session_id"])
-	})
-
-	t.Run("Success - returns session data without cookie", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/websession", nil)
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleWebSession(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.Equal(t, user.ID, resp["user_id"])
-		assert.Empty(t, resp["web_session_id"])
-	})
-}
-
-func TestHandleUsers(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/users", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleUsers(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/users", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleUsers(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "invalid JSON")
-	})
-
-	t.Run("Success - creates user", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"name": "Test User",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/users", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleUsers(rr, req)
-
-		assert.Equal(t, http.StatusCreated, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["user_id"])
-	})
-}
-
-func TestHandlePublicAuthLoginVerify(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/login/verify", nil)
-		rr := httptest.NewRecorder()
-
-		c.handlePublicAuthLoginVerify(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/login/verify", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handlePublicAuthLoginVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "invalid JSON")
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/login/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handlePublicAuthLoginVerify(rr, req)
-
-		assert.Equal(t, http.StatusUnauthorized, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user not found")
-	})
-}
-
-func TestHandlePublicAuthLogout(t *testing.T) {
-	t.Run("Success - clears cookie", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/logout", nil)
-		req.AddCookie(&http.Cookie{Name: "g8e_session", Value: "test-session"})
-		rr := httptest.NewRecorder()
-
-		c.handlePublicAuthLogout(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		cookies := rr.Result().Cookies()
-		assert.Len(t, cookies, 1)
-		assert.Equal(t, "g8e_session", cookies[0].Name)
-		assert.Equal(t, -1, cookies[0].MaxAge)
-	})
-
-	t.Run("Success - no cookie present", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/logout", nil)
-		rr := httptest.NewRecorder()
-
-		c.handlePublicAuthLogout(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-	})
-}
-
-func TestHandleCLIPasskeyRegisterChallenge(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/cli-register/challenge", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-register/challenge", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "invalid JSON body")
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Failure - user not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"user_id":   "nonexistent-user",
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user not found")
-	})
-
-	t.Run("Failure - first-credential only with existing credentials", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		// Add a fake credential to simulate existing credentials
-		user.PasskeyCredentials = []models.PasskeyCredential{{ID: []byte("existing-cred")}}
-		updatedUser, err := json.Marshal(user)
-		require.NoError(t, err)
-		c.db.DocStore.DocSet("users", user.ID, updatedUser)
-
-		body := map[string]string{
-			"user_id":   user.ID,
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "first-credential registration only")
-	})
-
-	t.Run("Success - valid request with mTLS enrollment", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id":   user.ID,
-			"user_name": "test-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-register/challenge", bytes.NewReader(b))
-		req = req.WithContext(context.WithValue(req.Context(), constants.ContextKeyUserID, user.ID))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["options"])
-	})
-}
-
-func TestHandleCLIPasskeyRegisterVerify(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/cli-register/verify", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-register/verify", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "invalid JSON body")
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-register/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Failure - user not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"user_id": "nonexistent-user",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-register/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user not found")
-	})
-
-	t.Run("Failure - first-credential only with existing credentials", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		// Add a fake credential to simulate existing credentials
-		user.PasskeyCredentials = []models.PasskeyCredential{{ID: []byte("existing-cred")}}
-		updatedUser, err := json.Marshal(user)
-		require.NoError(t, err)
-		c.db.DocStore.DocSet("users", user.ID, updatedUser)
-
-		body := map[string]string{
-			"user_id": user.ID,
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-register/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "first-credential registration only")
-	})
-}
-
-func TestHandleCLIBrowserPasskeyRegisterChallenge(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/cli-browser-register/challenge", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-browser-register/challenge", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "invalid JSON body")
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"user_name":     "test-user",
-			"cli_session_id": "session-123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-browser-register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Failure - user not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"user_id":        "nonexistent-user",
-			"user_name":      "test-user",
-			"cli_session_id": "session-123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-browser-register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user not found")
-	})
-
-	t.Run("Failure - first-credential only with existing credentials", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		// Add a fake credential to simulate existing credentials
-		user.PasskeyCredentials = []models.PasskeyCredential{{ID: []byte("existing-cred")}}
-		updatedUser, err := json.Marshal(user)
-		require.NoError(t, err)
-		c.db.DocStore.DocSet("users", user.ID, updatedUser)
-
-		body := map[string]string{
-			"user_id":        user.ID,
-			"user_name":      "test-user",
-			"cli_session_id": "session-123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-browser-register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "first-credential registration only")
-	})
-
-	t.Run("Success - valid request", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id":        user.ID,
-			"user_name":      "test-user",
-			"cli_session_id": "session-123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-browser-register/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterChallenge(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["options"])
-	})
-}
-
-func TestHandleCLIBrowserPasskeyRegisterVerify(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/cli-browser-register/verify", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-browser-register/verify", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "invalid JSON body")
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"cli_session_id": "session-123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-browser-register/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Failure - user not found", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"user_id":        "nonexistent-user",
-			"cli_session_id": "session-123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-browser-register/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user not found")
-	})
-
-	t.Run("Failure - first-credential only with existing credentials", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		// Add a fake credential to simulate existing credentials
-		user.PasskeyCredentials = []models.PasskeyCredential{{ID: []byte("existing-cred")}}
-		updatedUser, err := json.Marshal(user)
-		require.NoError(t, err)
-		c.db.DocStore.DocSet("users", user.ID, updatedUser)
-
-		body := map[string]string{
-			"user_id":        user.ID,
-			"cli_session_id": "session-123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli-browser-register/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIBrowserPasskeyRegisterVerify(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "first-credential registration only")
-	})
-}
-
-func TestHandleCLIPasskeyAuthenticateChallenge(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/cli/authenticate/challenge", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli/authenticate/challenge", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "invalid JSON body")
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli/authenticate/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-
-	t.Run("Failure - no passkeys registered", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		body := map[string]string{
-			"user_id": user.ID,
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli/authenticate/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.False(t, resp["success"].(bool))
-		assert.Contains(t, resp["error"].(string), "no passkeys registered")
-	})
-
-	t.Run("Success - valid request", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		user, err := c.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		// Add a fake credential to simulate existing credentials
-		user.PasskeyCredentials = []models.PasskeyCredential{{ID: []byte("existing-cred")}}
-		updatedUser, err := json.Marshal(user)
-		require.NoError(t, err)
-		c.db.DocStore.DocSet("users", user.ID, updatedUser)
-
-		body := map[string]string{
-			"user_id": user.ID,
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli/authenticate/challenge", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyAuthenticateChallenge(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["options"])
-	})
-}
-
-func TestHandleCLIPasskeyAuthenticateVerify(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/passkeys/cli/authenticate/verify", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyAuthenticateVerify(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli/authenticate/verify", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyAuthenticateVerify(rr, req)
+		reader := &fileActuatorKeyReader{path: keyFile}
+		keyID, publicKey, err := reader.ReadActuatorPublicKey()
 
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "invalid JSON body")
-	})
-
-	t.Run("Failure - missing user_id", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/passkeys/cli/authenticate/verify", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleCLIPasskeyAuthenticateVerify(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "user_id required")
-	})
-}
-
-func TestHandleDeviceEnrollment(t *testing.T) {
-	t.Run("Failure - method not allowed", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/auth/device/enroll", nil)
-		rr := httptest.NewRecorder()
-
-		c.handleDeviceEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Failure - invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/device/enroll", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-
-		c.handleDeviceEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "invalid JSON")
-	})
-
-	t.Run("Failure - missing csr_pem", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"system_fingerprint": "fp-123",
-			"hostname":          "test-host",
-		}
-		b, err := json.Marshal(body)
 		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/device/enroll", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleDeviceEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "csr_pem is required")
+		assert.Equal(t, "test-key-id", keyID)
+		assert.Equal(t, "test-public-key", publicKey)
 	})
 
-	t.Run("Failure - missing system_fingerprint", func(t *testing.T) {
+	t.Run("Failure - file does not exist", func(t *testing.T) {
 		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"csr_pem":   testutil.GenerateTestCSRP256(t, "test-device"),
-			"hostname": "test-host",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/device/enroll", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
+		reader := &fileActuatorKeyReader{path: "/nonexistent/path/actuator_key.json"}
+		_, _, err := reader.ReadActuatorPublicKey()
 
-		c.handleDeviceEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "system_fingerprint is required")
+		assert.Error(t, err)
+		assert.True(t, os.IsNotExist(err))
 	})
 
-	t.Run("Failure - missing hostname", func(t *testing.T) {
+	t.Run("Failure - invalid JSON in file", func(t *testing.T) {
 		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		body := map[string]string{
-			"csr_pem":            testutil.GenerateTestCSRP256(t, "test-device"),
-			"system_fingerprint": "fp-123",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/device/enroll", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
+		tmpDir := t.TempDir()
+		keyFile := filepath.Join(tmpDir, "actuator_key.json")
+		require.NoError(t, os.WriteFile(keyFile, []byte("{invalid json"), 0644))
 
-		c.handleDeviceEnrollment(rr, req)
+		reader := &fileActuatorKeyReader{path: keyFile}
+		_, _, err := reader.ReadActuatorPublicKey()
 
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.Contains(t, rr.Body.String(), "hostname is required")
+		assert.Error(t, err)
 	})
 
-	t.Run("Failure - bootstrap user disabled", func(t *testing.T) {
+	t.Run("Success - missing required fields returns empty values", func(t *testing.T) {
 		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		bootstrapUser, err := c.userSvc.CreateBootstrapUser()
-		require.NoError(t, err)
-		c.userSvc.Disable(bootstrapUser.ID, "test", "actor", "op")
-
-		body := map[string]string{
-			"csr_pem":            testutil.GenerateTestCSRP256(t, "test-device"),
-			"system_fingerprint": "fp-123",
-			"hostname":          "test-host",
-		}
-		b, err := json.Marshal(body)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/device/enroll", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleDeviceEnrollment(rr, req)
+		tmpDir := t.TempDir()
+		keyFile := filepath.Join(tmpDir, "actuator_key.json")
+		require.NoError(t, os.WriteFile(keyFile, []byte(`{"key_id":"test-id"}`), 0644))
 
-		assert.Equal(t, http.StatusConflict, rr.Code)
-		assert.Contains(t, rr.Body.String(), "bootstrap user is disabled")
-	})
-
-	t.Run("Failure - device enrollment on non-empty system", func(t *testing.T) {
-		t.Parallel()
-		c, _ := setupTestAuthController(t)
-		// Create a regular user to make system non-empty
-		c.userSvc.CreateUser()
+		reader := &fileActuatorKeyReader{path: keyFile}
+		keyID, publicKey, err := reader.ReadActuatorPublicKey()
 
-		body := map[string]string{
-			"csr_pem":            testutil.GenerateTestCSRP256(t, "test-device"),
-			"system_fingerprint": "fp-123",
-			"hostname":          "test-host",
-		}
-		b, err := json.Marshal(body)
 		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/auth/device/enroll", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		c.handleDeviceEnrollment(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.Contains(t, rr.Body.String(), "device enrollment only available for initial setup")
+		assert.Equal(t, "test-id", keyID)
+		assert.Empty(t, publicKey) // Missing field returns empty string
 	})
 }
