+++ internal/services/gateway/state_root_service_test.go
@@ -134,3 +134,65 @@ func TestStateRootService_CalculateStateRootUncached(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, root, persistedRoot)
 }
+
+// TestStateRootService_NoCacheLeakOnDocumentWrite verifies that inserting a document
+// and then calling GetCurrentStateRoot does NOT change the returned root between the
+// insert and the root call (cache invalidation must not leak into the authoritative root).
+// This is a regression test for the bug where cache-key churn in kv_store caused
+// state root mismatches.
+func TestStateRootService_NoCacheLeakOnDocumentWrite(t *testing.T) {
+	t.Parallel()
+	db := newTestDB(t)
+	svc := NewStateRootService(db.GetDB(), testutil.NewTestLogger())
+
+	// Get initial state root
+	root1, err := svc.GetCurrentStateRoot()
+	require.NoError(t, err)
+	assert.NotEmpty(t, root1)
+
+	// Add a cache entry (simulating cache invalidation)
+	err = db.KVStore.KVSet("g8e:cache:doc:test:doc1", "cached_value", 3600)
+	require.NoError(t, err)
+
+	// Get state root - should NOT change because cache keys are excluded
+	root2, err := svc.GetCurrentStateRoot()
+	require.NoError(t, err)
+	assert.Equal(t, root1, root2, "state root should not change when cache keys are added")
+
+	// Add another cache entry with different prefix
+	err = db.KVStore.KVSet("g8e:cache:query:SELECT * FROM test", "query_result", 3600)
+	require.NoError(t, err)
+
+	// Get state root - should still not change
+	root3, err := svc.GetCurrentStateRoot()
+	require.NoError(t, err)
+	assert.Equal(t, root2, root3, "state root should not change when more cache keys are added")
+
+	// Add a non-cache KV entry (authoritative state)
+	err = db.KVStore.KVSet("authoritative:key", "value", 0)
+	require.NoError(t, err)
+
+	// Get state root - should now change
+	root4, err := svc.GetCurrentStateRoot()
+	require.NoError(t, err)
+	assert.NotEqual(t, root3, root4, "state root should change when authoritative KV entries are added")
+
+	// Add a document (which triggers cache invalidation internally)
+	docData := mustDocJSON(t, map[string]interface{}{"key": "value"})
+	err = db.DocStore.DocSet("test", "doc1", docData)
+	require.NoError(t, err)
+
+	// Get state root - should change due to document, not due to cache churn
+	root5, err := svc.GetCurrentStateRoot()
+	require.NoError(t, err)
+	assert.NotEqual(t, root4, root5, "state root should change after document insertion")
+
+	// Add another document to verify consistent behavior
+	docData2 := mustDocJSON(t, map[string]interface{}{"key2": "value2"})
+	err = db.DocStore.DocSet("test", "doc2", docData2)
+	require.NoError(t, err)
+
+	root6, err := svc.GetCurrentStateRoot()
+	require.NoError(t, err)
+	assert.NotEqual(t, root5, root6, "state root should change after second document insertion")
+}
