+++ test/scenario/scenario_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build e2e
+//go:build integration
 
 package scenario
 
@@ -19,20 +19,19 @@ import (
 	"context"
 	"crypto/ed25519"
 	"encoding/json"
+	"encoding/pem"
 	"fmt"
 	"net/http"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 
 	"github.com/g8e-ai/g8e/internal/cli/auth"
-	cliconfig "github.com/g8e-ai/g8e/internal/cli/config"
+	"github.com/g8e-ai/g8e/internal/config"
 	"github.com/g8e-ai/g8e/internal/constants"
 	"github.com/g8e-ai/g8e/internal/emulator/client"
-	"github.com/g8e-ai/g8e/internal/emulator/config"
+	emulatorconfig "github.com/g8e-ai/g8e/internal/emulator/config"
+	"github.com/g8e-ai/g8e/test/fixtures"
 	commonv1 "github.com/g8e-ai/g8e/protocol/proto/g8e/common/v1"
 	operatorv1 "github.com/g8e-ai/g8e/protocol/proto/g8e/operator/v1"
 	"google.golang.org/protobuf/encoding/protojson"
@@ -55,173 +54,66 @@ type TestContext struct {
 	PubKey            ed25519.PublicKey
 	OperatorSessionID string
 	CLISessionID      string
+	Fixture           *fixtures.GatewayFixture
+	Identity          *fixtures.ClientIdentity
 }
 
-// setupTestContext connects to a real running operator/gateway.
+// setupTestContext spins up an in-process gateway via GatewayFixture
+// and enrolls a client identity for mTLS authentication.
 // Returns a TestContext with mTLS client ready for use.
 func setupTestContext(t *testing.T) *TestContext {
 	t.Helper()
 
-	// Initialize paths relative to project root using constant
-	if err := constants.InitPathsWithBase(constants.ProjectRootFromTestDir); err != nil {
-		t.Fatalf("failed to initialize paths: %v", err)
-	}
+	// Create in-process gateway
+	f := fixtures.NewGatewayFixture(t, fixtures.GatewayFixtureOptions{
+		TestName: "scenario-test",
+		Posture:  config.PostureDoctrine,
+	})
 
-	repoRoot := constants.ProjectRootFromTestDir
-	projectRoot := constants.Paths.Infra.RuntimeDir
+	// Enroll a client identity for mTLS authentication
+	identity := fixtures.EnrollClientIdentity(t, f, "scenario-user", "scenario-org", "scenario-fingerprint", "scenario-host")
 
-	// Ensure platform is running and authenticated
-	// We build the g8e binary if needed
-	g8ePath := filepath.Join(repoRoot, "g8e")
-	if _, err := os.Stat(g8ePath); os.IsNotExist(err) {
-		buildCmd := exec.Command("go", "build", "-o", g8ePath, "./cmd/operator")
-		buildCmd.Dir = repoRoot
-		if output, err := buildCmd.CombinedOutput(); err != nil {
-			t.Fatalf("failed to build g8e binary: %s", string(output))
-		}
-	}
-
-	testEnv := os.Getenv("G8E_TEST_ENV")
-	useDocker := testEnv == "docker" || strings.HasPrefix(testEnv, "demos/")
-
-	if useDocker {
-		composeFile := "docker-compose.yml"
-		if strings.HasPrefix(testEnv, "demos/") {
-			composeFile = filepath.Join(testEnv, "compose.yml")
-		}
-		composePath := filepath.Join(repoRoot, composeFile)
-
-		t.Logf("Ensuring Docker environment is up (using %s)...", composeFile)
-		upCmd := exec.Command("docker", "compose", "-f", composePath, "up", "-d")
-		upCmd.Dir = repoRoot
-		if output, err := upCmd.CombinedOutput(); err != nil {
-			t.Fatalf("failed to start docker-compose: %s", string(output))
-		}
-	} else {
-		// Local start if not already running
-		checkCmd := exec.Command(g8ePath, "gw", "status")
-		checkCmd.Dir = repoRoot
-		checkOutput, _ := checkCmd.CombinedOutput()
-		if strings.Contains(string(checkOutput), "STOPPED") {
-			t.Logf("Gateway not running, starting it...")
-			startCmd := exec.Command(g8ePath, "gw", "start")
-			startCmd.Dir = repoRoot
-			if output, err := startCmd.CombinedOutput(); err != nil {
-				t.Fatalf("failed to start gateway: %s", string(output))
-			}
-		}
-	}
-
-	cliCfg, err := cliconfig.Load(projectRoot)
+	// Write certificates to temp files for emulator client
+	certFile, err := os.CreateTemp("", "scenario-cert-*.pem")
 	if err != nil {
-		t.Fatalf("failed to load CLI config: %v", err)
+		t.Fatalf("failed to create temp cert file: %v", err)
 	}
-
-	// Bootstrap CLI authentication if not already done
-	credsPath := filepath.Join(projectRoot, ".g8e", "credentials")
-	needsLogin := true
-	if info, statErr := os.Stat(credsPath); statErr == nil {
-		if time.Since(info.ModTime()) < 45*time.Minute {
-			needsLogin = false
-		}
+	defer os.Remove(certFile.Name())
+	if _, err := certFile.Write(identity.Certificate); err != nil {
+		t.Fatalf("failed to write cert file: %v", err)
 	}
+	certFile.Close()
 
-	if needsLogin {
-		t.Logf("Performing non-interactive auth login...")
-		loginCmd := exec.Command(g8ePath, "auth", "login")
-		loginCmd.Dir = repoRoot
-		loginCmd.Env = append(os.Environ(), "G8E_SKIP_PASSKEY=true")
-		if output, err := loginCmd.CombinedOutput(); err != nil {
-			t.Fatalf("failed to run './g8e auth login': %s", string(output))
-		}
+	keyFile, err := os.CreateTemp("", "scenario-key-*.pem")
+	if err != nil {
+		t.Fatalf("failed to create temp key file: %v", err)
 	}
-
-	// Ensure gateway is running and governance is ready before proceeding
-	healthURL := fmt.Sprintf("http://127.0.0.1:%d/api/v1/health", constants.Ports.OperatorHttp)
-	t.Logf("Waiting for gateway to be governance-ready at %s...", healthURL)
-
-	timeout := 5 * time.Second
-	if useDocker {
-		timeout = 60 * time.Second
+	defer os.Remove(keyFile.Name())
+	keyPEM := pem.EncodeToMemory(&pem.Block{Type: "EC PRIVATE KEY", Bytes: identity.PrivateKey})
+	if _, err := keyFile.Write(keyPEM); err != nil {
+		t.Fatalf("failed to write key file: %v", err)
 	}
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		resp, err := http.Get(healthURL)
-		if err != nil {
-			t.Logf("Gateway not ready yet, retrying... (%v)", err)
-			time.Sleep(500 * time.Millisecond)
-			continue
-		}
-		defer resp.Body.Close()
+	keyFile.Close()
 
-		if resp.StatusCode != http.StatusOK {
-			t.Logf("Gateway returned status %d, retrying...", resp.StatusCode)
-			time.Sleep(500 * time.Millisecond)
-			continue
-		}
+	// Read CA bundle from PKI dir
+	caBundlePath := f.PKIDir + "/trust/g8eg-ca-bundle.pem"
 
-		var health struct {
-			Status          string `json:"status"`
-			GovernanceReady bool   `json:"governance_ready"`
-		}
-		if err := json.NewDecoder(resp.Body).Decode(&health); err != nil {
-			t.Logf("Failed to decode health response, retrying... (%v)", err)
-			time.Sleep(500 * time.Millisecond)
-			continue
-		}
-
-		if health.GovernanceReady {
-			t.Logf("Gateway is governance-ready")
-			break
-		}
-
-		t.Logf("Gateway not governance-ready yet, retrying...")
-		time.Sleep(500 * time.Millisecond)
-	}
-
-	if time.Now().After(deadline) {
-		t.Fatal("gateway did not become governance-ready within timeout")
-	}
-
-	// Paths to local PKI material (bootstrapped via ./g8e auth login)
-	clientCertPath := cliCfg.CLICertFile()
-	clientKeyPath := cliCfg.CLIKeyFile()
-	caBundlePath := constants.Paths.Infra.CaCertPath
-
-	// Verify certificates exist
-	if _, err := os.Stat(clientCertPath); os.IsNotExist(err) {
-		t.Fatalf("client cert not found at %s", clientCertPath)
-	}
-	if _, err := os.Stat(caBundlePath); os.IsNotExist(err) {
-		t.Fatalf("CA bundle not found at %s", caBundlePath)
-	}
-
-	// Create auditor client for HTTP submission
-	auditorCfg := config.Default()
+	// Create emulator client for HTTP submission
+	mtlsURL := constants.LocalhostHTTPSURL(f.Service.GetHTTPSPort())
+	auditorCfg := emulatorconfig.Default()
 	auditorCfg.UseCLIConfig = false
-	auditorCfg.MTLSBaseURL = constants.LocalhostHTTPSURL(constants.Ports.OperatorHttps)
-	auditorCfg.PublicBaseURL = constants.LocalhostHTTPSURL(constants.Ports.OperatorHttps)
-	auditorCfg.Auth.ClientCert = clientCertPath
-	auditorCfg.Auth.ClientKey = clientKeyPath
+	auditorCfg.MTLSBaseURL = mtlsURL
+	auditorCfg.PublicBaseURL = mtlsURL
+	auditorCfg.Auth.ClientCert = certFile.Name()
+	auditorCfg.Auth.ClientKey = keyFile.Name()
 	auditorCfg.Auth.CABundle = caBundlePath
 	auditorCfg.Auth.Insecure = true
 	auditorCfg.Verbose = true
 
-	// Load CLI credentials
-	creds, err := auth.LoadCredentials(cliCfg)
-	if err != nil || creds == nil {
-		t.Fatalf("failed to load CLI credentials")
-	}
-
-	t.Logf("Creating auditor client with Cert: %s, Key: %s, CA: %s", clientCertPath, clientKeyPath, caBundlePath)
+	t.Logf("Creating auditor client with Cert: %s, Key: %s, CA: %s", certFile.Name(), keyFile.Name(), caBundlePath)
 	testClient, err := client.New(auditorCfg)
 	if err != nil {
 		t.Fatalf("failed to create auditor client: %v", err)
-	}
-
-	operatorSessionID := creds.CLISessionID
-	if operatorSessionID == "" {
-		t.Fatal("CLI session ID is empty")
 	}
 
 	pub, priv, err := ed25519.GenerateKey(nil)
@@ -231,41 +123,74 @@ func setupTestContext(t *testing.T) *TestContext {
 
 	return &TestContext{
 		Client:            testClient,
-		BaseURL:           auditorCfg.MTLSBaseURL,
-		CertPath:          clientCertPath,
-		KeyPath:           clientKeyPath,
+		BaseURL:           mtlsURL,
+		CertPath:          certFile.Name(),
+		KeyPath:           keyFile.Name(),
 		CAPath:            caBundlePath,
 		PrivKey:           priv,
 		PubKey:            pub,
-		OperatorSessionID: operatorSessionID,
-		CLISessionID:      creds.CLISessionID,
+		OperatorSessionID: identity.OperatorSessionID,
+		CLISessionID:      identity.OperatorSessionID,
+		Fixture:           f,
+		Identity:          identity,
 	}
+}
+
+func logStateVersion(t *testing.T, label string, f *fixtures.GatewayFixture) {
+	t.Helper()
+	db := f.Service.GetDB().GetDB()
+	var version int64
+	if err := db.QueryRowWithRetry("SELECT version FROM state_version WHERE id = 1").Scan(&version); err != nil {
+		t.Logf("state_version [%s]: query error: %v", label, err)
+		return
+	}
+	t.Logf("state_version [%s]: %d", label, version)
+	rows, err := db.QueryWithRetry("SELECT collection, id FROM documents ORDER BY collection, id")
+	if err == nil {
+		defer rows.Close()
+		for rows.Next() {
+			var col, id string
+			rows.Scan(&col, &id)
+			t.Logf("  doc: %s/%s", col, id)
+		}
+	}
+	var kvCount int
+	db.QueryRowWithRetry("SELECT COUNT(*) FROM kv_store").Scan(&kvCount)
+	t.Logf("  kv_store rows: %d", kvCount)
 }
 
 func TestScenarios(t *testing.T) {
 	// Setup test infrastructure
 	ctx := setupTestContext(t)
+	defer ctx.Fixture.Cleanup()
 
-	// Fetch actual state root from gateway via mTLS port
-	// Note: StateRoot uses PublicBaseURL by default, but in full cert mode
-	// all ports require mTLS, so we need to use the mTLS endpoint directly
-	stateRoot, err := ctx.Client.StateRootFromMTLS(context.Background())
+	logStateVersion(t, "after-setup", ctx.Fixture)
+
+	// Fetch actual state root directly from StateRootService
+	stateRoot, err := ctx.Fixture.Service.GetDB().StateRootSvc.GetCurrentStateRoot()
 	if err != nil {
 		t.Fatalf("failed to fetch state root: %v", err)
 	}
-	t.Logf("State root from gateway: %q", stateRoot)
+	t.Logf("State root from service: %q", stateRoot)
+
+	logStateVersion(t, "after-state-root-fetch", ctx.Fixture)
 
 	// Build a valid envelope using the builder
 	intentBytes, err := New().
 		WithCommand("echo hello").
-		WithOperatorID("").
+		WithOperatorID(ctx.Identity.OperatorID).
 		WithOperatorSessionID(ctx.OperatorSessionID).
 		WithStateRoot(stateRoot).
-		WithL2(ctx.PrivKey, true).
 		Build()
 	if err != nil {
 		t.Fatalf("failed to build test envelope: %v", err)
 	}
+
+	// Debug: check state root directly before submission
+	directRoot, directErr := ctx.Fixture.Service.GetDB().StateRootSvc.GetCurrentStateRoot()
+	t.Logf("Direct state root before submit: %q (err=%v)", directRoot, directErr)
+	t.Logf("Envelope state root:             %q", stateRoot)
+	t.Logf("Roots match: %v", directRoot == stateRoot)
 
 	// Submit via real HTTP client
 	creds := &auth.Credentials{
@@ -305,6 +230,7 @@ func TestScenarios(t *testing.T) {
 // malformed envelopes are correctly rejected.
 func TestNegativeControls(t *testing.T) {
 	ctx := setupTestContext(t)
+	defer ctx.Fixture.Cleanup()
 
 	t.Run("bad_id_rejection", func(t *testing.T) {
 		intentBytes, err := New().
